♻️ refactor: 기존 product.py 삭제 및 import 경로 정리 완료
- 기존 monolithic product.py 파일 삭제
- views/__init__.py를 새로운 패키지 구조에 맞게 업데이트
- MonthlyFeaturedDrinksView 추가로 완전한 메인페이지 구성
- Product 뷰 리팩토링 완전 완료
