+++ apps/products/views/__init__.py
@@ -1,4 +1,4 @@
-# apps/products/views/__init__.py 수정
+# apps/products/views/__init__.py
 
 from .brewery import (
     BreweryCreateView,
@@ -9,16 +9,31 @@ from .brewery import (
 from .drink import (
     DrinkListView,
 )
+
+# 새로운 product 패키지 구조에서 import
 from .product import (
-    AwardWinningProductsView,
-    FeaturedProductsView,
-    MakgeolliProductsView,
-    PopularProductsView,
+    # 일반 사용자용 API
+    ProductSearchView,
     ProductDetailView,
     ProductLikeToggleView,
-    ProductSearchView,
+
+    # 메인페이지 섹션들
+    MonthlyFeaturedDrinksView,
+    PopularProductsView,
     RecommendedProductsView,
+
+    # 패키지페이지 섹션들
+    FeaturedProductsView,
+    AwardWinningProductsView,
+    MakgeolliProductsView,
     RegionalProductsView,
+
+    # 관리자용 API (필요한 경우)
+    IndividualProductCreateView,
+    PackageProductCreateView,
+    DrinksForPackageView,
+    ProductManageView,
+    ProductManageListView,
 )
 
 __all__ = [
@@ -27,16 +42,30 @@ __all__ = [
     "BreweryDetailView",
     "BreweryCreateView",
     "BreweryManageView",
+
     # Drink (관리자용 - 목록만)
     "DrinkListView",
-    # Product (UI용)
+
+    # Product - 일반 사용자용 API
     "ProductSearchView",
     "ProductDetailView",
     "ProductLikeToggleView",
+
+    # Product - 메인페이지 섹션들
+    "MonthlyFeaturedDrinksView",
     "PopularProductsView",
+    "RecommendedProductsView",
+
+    # Product - 패키지페이지 섹션들
     "FeaturedProductsView",
     "AwardWinningProductsView",
     "MakgeolliProductsView",
     "RegionalProductsView",
-    "RecommendedProductsView",
+
+    # Product - 관리자용 API
+    "IndividualProductCreateView",
+    "PackageProductCreateView",
+    "DrinksForPackageView",
+    "ProductManageView",
+    "ProductManageListView",
 ]
